% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{calc_comp_rel}
\alias{calc_comp_rel}
\title{Calculate the composite reliability}
\usage{
calc_comp_rel(fit)
}
\arguments{
\item{fit}{A fitted factor model.}
}
\value{
A numeric vector indicating the omega reliability of all the latent
factors.
}
\description{
Note based on Flora (2020), we calculated omega reliability as the variance
explained by g factor.
}
