% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_mask.R
\name{calc_mask_dice}
\alias{calc_mask_dice}
\title{Calculate Dice coefficient for brain mask}
\usage{
calc_mask_dice(mask, ...)
}
\arguments{
\item{mask}{A vector of brain mask, should be the upper triangle of the
original full mask.}

\item{...}{Other arguments passed to \code{\link[=binarize_mask]{binarize_mask()}}.}
}
\value{
A numeric value of Dice coefficient.
}
\description{
Calculate Dice coefficient for brain mask
}
