% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scoring.R
\name{calc_var_exp}
\alias{calc_var_exp}
\title{Calculate the variance explained by g factor}
\usage{
calc_var_exp(fit)
}
\arguments{
\item{fit}{A fitted one-g factor model.}
}
\value{
A numeric value indicating the variance explained by g factor.
}
\description{
Calculate the variance explained by g factor
}
