% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{combine_targets}
\alias{combine_targets}
\title{Target factory to combine targets from different batches}
\usage{
combine_targets(name, targets, cols_targets)
}
\arguments{
\item{name}{The name for the target. Should be a symbol.}

\item{targets}{A list from which targets to extract.}

\item{cols_targets}{A character vector specifying the columns in the names of
targets.}
}
\value{
A new target object to do the given targets combination.
}
\description{
Target factory to combine targets from different batches
}
