% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{correct_subjs_id}
\alias{correct_subjs_id}
\title{Correct Subjects' identifier}
\usage{
correct_subjs_id(data, sub_id_transform, name_id = "sub_id")
}
\description{
Some subjects have different identifier in behavior and fmri data. Here we
correct them as fmri identifiers.
}
