% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{cpm2}
\alias{cpm2}
\title{Perform Connectome-based Predictive Modeling}
\usage{
cpm2(
  data,
  behav = NULL,
  kfolds = NULL,
  bias_correct = NULL,
  thresh_method = c("alpha", "sparsity"),
  thresh_level = 0.01
)
}
\arguments{
\item{data}{A matrix contains connectome data. Observations in row, edges in
column (stretch upper triangular matrix).}

\item{behav}{A numeric vector contains behavior data. Length must equal to
number of observations in \code{data}. If \code{NULL}, the last column of \code{data} will
be extracted to be \code{behav}.}

\item{kfolds}{Folds number of cross-validation. If \code{NULL}, it will be set to
be equal to the number of observations, i.e., leave-one-subject-out.}

\item{bias_correct}{Logical value indicating if z-score should be performed
separated in training and test group. This is forced to be \code{FALSE} for
leave one out case. Otherwise, it is \code{TRUE} by default.}

\item{thresh_method, thresh_level}{The threshold method and level used in edge
selection. Choices are correlation p value (\code{"alpha"}) and network sparsity
(\code{"sparsity"}) methods.}
}
\value{
A list contains following fields:
\itemize{
\item mask_prop_{pos,neg}: Positive mask and negative mask, values are the
selection proportion.
\item behav_pred_{pos,neg,all}: Predicted behavior for positive, negative and
all edges.
\item cor_{pos,neg,all}: Correlation test result between predicted and true
behavior.
}
}
\description{
This is just a single run of the whole protocol (not including permutation).
The number \code{2} in the function name is initially used to distinguish from the
\code{cpm()} which is now removed, but the name will be kept for historical
reasons.
}
\author{
Liang Zhang \href{mailto:psychelzh@outlook.com}{psychelzh@outlook.com}
}
