% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{do_cpm}
\alias{do_cpm}
\title{These functions are all deprecated for certain reasons. Please do not use
them now.}
\usage{
do_cpm(fc_data, scores, thresh_method, thresh_level)
}
\description{
These functions are all deprecated for certain reasons. Please do not use
them now.
}
