% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cpm.R
\name{do_cpm2}
\alias{do_cpm2}
\title{Perform CPM by correctly join neural and behavioral data}
\usage{
do_cpm2(
  neural,
  behav,
  kfolds,
  thresh_method,
  thresh_level,
  bias_correct = TRUE,
  id_cols = "sub_id"
)
}
\description{
Perform CPM by correctly join neural and behavioral data
}
