% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{extract_brain_mask}
\alias{extract_brain_mask}
\title{Extract CPM mask results}
\usage{
extract_brain_mask(result_cpm, by, col_cpm = cpm)
}
\arguments{
\item{result_cpm}{A list of results from \code{cpm2()}.}

\item{by}{A character vector specifying the column names for grouping
variables.}

\item{col_cpm}{A character vector specifying the column names for CPM
prediction results.}
}
\value{
A data frame with CPM mask results.
}
\description{
Extract CPM mask results
}
