% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cpm.R
\name{extract_cpm_pred}
\alias{extract_cpm_pred}
\title{Extract CPM prediction results}
\usage{
extract_cpm_pred(result_cpm, col_cpm = cpm)
}
\arguments{
\item{result_cpm}{A list of results from \code{cpm2()}.}

\item{col_cpm}{A character vector specifying the column names for CPM
prediction results.}
}
\value{
A data frame with CPM prediction results.
}
\description{
Extract CPM prediction results
}
