% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_g}
\alias{fit_g}
\title{Fit a one-g factor model for given observed variables}
\usage{
fit_g(data, vars)
}
\arguments{
\item{data}{Raw behavior data.}

\item{vars}{A character vector specifying observed variables.}
}
\value{
A fitted one-g factor model.
}
\description{
Fit a one-g factor model for given observed variables
}
