% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{permute_cpm2}
\alias{permute_cpm2}
\title{Target factory for CPM permutation}
\usage{
permute_cpm2(
  behav,
  config_neural,
  hypers_cpm,
  subjs_subset = NULL,
  name_suffix = "",
  include_file_targets = TRUE,
  split_hyper = NULL,
  subjs_info = NULL,
  batches = 4,
  reps = 5
)
}
\arguments{
\item{behav}{The name for the behavioral data. Should be a symbol.}

\item{config_neural}{A \code{\link[=data.frame]{data.frame()}} storing the specifications of neural
data used. The \code{tar_neural} and \code{file} fields must be present to specify
which neural data to use.}

\item{hypers_cpm}{A \code{\link[=data.frame]{data.frame()}} storing the CPM hyper parameters passed to
the \code{values} argument of \code{\link[tarchetypes:tar_map_rep]{tarchetypes::tar_map_rep()}}. Note the names must
be consistent with the names of \code{formals(do_cpm2)}.}

\item{subjs_subset}{The subject list to include in CPM analysis.}

\item{name_suffix}{The name suffix for the CPM targets.}

\item{split_hyper, subjs_info}{If one of these two parameters is specified,
the other must be specified, too. \code{split_hyper} specifies the field used to
split neural data to perform different CPM calculations, e.g., different
gender. Note this field must be present in both \code{hypers} and \code{subjs_info}.}

\item{batches, reps}{The number of batches and repetitions passed to
\code{\link[tarchetypes:tar_map_rep]{tarchetypes::tar_map_rep()}}.}
}
\value{
A list of new target objects to calculate the permutation results.
}
\description{
This will generate batches of CPM permutation for targets to use.
}
