% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit-g.R
\name{predict_g_score}
\alias{predict_g_score}
\title{Predict g factor scores}
\usage{
predict_g_score(data, mdl, id_cols = 1)
}
\arguments{
\item{data}{Raw behavior data.}

\item{mdl}{A fitted one-g factor model.}

\item{id_cols}{A numeric vector specifying the column indices of subject
identifiers.}
}
\value{
A data frame with g factor scores.
}
\description{
Predict g factor scores
}
