% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{prepare_neural_files}
\alias{prepare_neural_files}
\title{Let targets track all the required neural data}
\usage{
prepare_neural_files(config, name_prefix = NULL, names = NULL)
}
\arguments{
\item{config}{Require \code{"modal"}, \code{"parcel"} and \code{"gsr"}.}

\item{name_prefix}{The targets name prefix. Must be a scalar character.}

\item{name}{description}
}
\description{
Let targets track all the required neural data
}
