% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{prepare_permute_cpm2}
\alias{prepare_permute_cpm2}
\title{Target factory for CPM permutation}
\usage{
prepare_permute_cpm2(
  config_neural,
  hypers_cpm = NULL,
  behav = NULL,
  subjs_subset = NULL,
  name_suffix = "",
  include_file_targets = TRUE,
  split_hyper = NULL,
  subjs_info = NULL,
  batches = 4,
  reps = 5
)
}
\arguments{
\item{config_neural}{A \code{\link[=data.frame]{data.frame()}} storing the specifications of neural
data used. The \code{tar_neural} and \code{file} fields must be present to specify
which neural data to use.}

\item{hypers_cpm}{A \code{\link[=data.frame]{data.frame()}} storing the CPM hyper parameters passed to
the \code{values} argument of \code{\link[tarchetypes:tar_map_rep]{tarchetypes::tar_map_rep()}}. Note the names must
be consistent with the argument names of \code{\link[=do_cpm2]{do_cpm2()}}. If this argument is
not specified or is \code{NULL}, it will only return targets tracking the neural
data.}

\item{behav}{The expression to get behavioral data.}

\item{subjs_subset}{The expression to get the subject list to include in CPM
analysis.}

\item{name_suffix}{A character scalar specifying the name suffix for the CPM
targets.}

\item{split_hyper, subjs_info}{If one of these two parameters is specified,
the other must be specified, too. \code{split_hyper} should be a character
scalar specifying the field used to split neural data to perform different
CPM calculations, e.g., different gender. Note this field must be present
in both \code{hypers_cpm} and \code{subjs_info}. \code{subjs_info} is an expression to get
the required subjects' information used to filter out corresponding data to
do CPM calculations.}

\item{batches, reps}{The number of batches and repetitions passed to
\code{\link[tarchetypes:tar_map_rep]{tarchetypes::tar_map_rep()}}.}
}
\value{
A list of new target objects to calculate the permutation results.
If \code{hypers_cpm} is the not specified or is \code{NULL}, a list of targets
tracking neural data will be returned.
}
\description{
This will generate batches of CPM permutation for targets to use.
}
