% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{prepare_roi_labels}
\alias{prepare_roi_labels}
\title{Prepare ROI labels}
\usage{
prepare_roi_labels(atlas, ...)
}
\arguments{
\item{atlas}{A data frame contains ROI labels and colors of a specific atlas.}

\item{...}{For future usage. Should be empty.}
}
\value{
A data frame of ROI labels. Note that the \code{label_hemi} column is
added as a combination of \code{hemi} and \code{label} columns, separated by \code{"_"}.
}
\description{
Prepare ROI labels
}
