% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scoring.R
\name{regress_covariates}
\alias{regress_covariates}
\title{Regress out covariates}
\usage{
regress_covariates(data, subjs_info, covars = NULL)
}
\arguments{
\item{data}{A data frame with subject identifiers and outcome variables.}

\item{subjs_info}{A data frame with subject identifiers and covariates.}

\item{covars}{A character vector specifying covariates. If \code{NULL}, all
covariates will be included.}
}
\value{
A data frame with residuals.
}
\description{
Regress out covariates
}
