% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{split_data_solomon}
\alias{split_data_solomon}
\title{Split data into two equivalent halves by SOLOMON method}
\usage{
split_data_solomon(data, id_cols = "sub_id")
}
\arguments{
\item{data}{The data to split.}

\item{id_cols}{The names for identifier columns.}
}
\value{
A list of two equivalent data.
}
\description{
Details can be found at Lorenzo-Seva, 2021.
}
