% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{summarise_adjacency}
\alias{summarise_adjacency}
\title{Summarise adjacency matrix}
\usage{
summarise_adjacency(adj_mat, labels)
}
\arguments{
\item{adj_mat}{A matrix of adjacency.}

\item{labels}{A vector of labels. Must have the same length as the number of
rows/columns of the adjacency matrix.}
}
\description{
This will add labels to the adjacency matrix and summarise the adjacency
matrix to get the degree, number of links, relative degree and contribution
of each ROI.
}
