% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_chord}
\alias{visualize_chord}
\title{Visualize brain mask as chord diagram}
\usage{
visualize_chord(
  adj_mat,
  roi_info,
  ...,
  model_type = NULL,
  link_val = c("n", "total", "prop", "enrich"),
  link_color = c("white", "black"),
  group_by_hemi = TRUE,
  thresh = NULL,
  add_legend = FALSE
)
}
\arguments{
\item{adj_mat}{A matrix of adjacency.}

\item{roi_info}{A data frame contains ROI labels and colors.}

\item{...}{Further arguments passed to \code{\link[=visualize_chord]{visualize_chord()}}. Currently not
used.}

\item{model_type}{A character string specifying the type of the edge. Can be
\code{"pos"} or \code{"neg"}. Note that this argument is exclusive with \code{link_color}.}

\item{link_val}{A character string specifying the value of the link.}

\item{link_color}{A character string of length two, specifying the color of
the lowest and highest link value. If \code{model_type} is specified, this
argument should not be used, and the color will be set automatically.}

\item{group_by_hemi}{A logical value indicating whether to group ROIs by
hemisphere.}

\item{thresh}{A numeric value specifying the threshold of the value to be
visualized. If the value is less than the threshold, it will be set as
0.}

\item{add_legend}{A logical value indicating whether to add a legend.}
}
\value{
Invisible \code{NULL}.
}
\description{
Visualize brain mask as chord diagram
}
