% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_mask.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize brain mask as chord diagram}
\usage{
visualize_network(
  adj_df,
  roi_labels,
  link_val = c("degree", "prop"),
  link_color = c("white", "black")
)
}
\arguments{
\item{adj_df}{A data frame contains adjacency matrix, should be the upper
triangle of the original full matrix.}

\item{roi_labels}{A data frame contains ROI labels and colors.}

\item{link_val}{A character string of link value, should be one of \code{"degree"}
and \code{"prop"}.}

\item{link_color}{A character string of length two, specifying the color of
the lowest and highest link value.}
}
\value{
Invisible \code{NULL}.
}
\description{
Visualize brain mask as chord diagram
}
