% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_mask.R
\name{binarize_mask}
\alias{binarize_mask}
\title{Binarize brain mask}
\usage{
binarize_mask(
  mask,
  binarize_method = c("prob", "count"),
  binarize_level = NULL
)
}
\arguments{
\item{mask}{A vector of brain mask, should be the upper triangle of the
original full mask.}

\item{binarize_method}{A character string of binarization method, should be
one of \code{"prob"} and \code{"count"}.}

\item{binarize_level}{A numeric value of binarization level, should be a
number between 0 and 1 (default: \code{0.95}) if \code{binarize_method} is \code{"prob"},
or a number between 1 and length of \code{mask} (default: \code{500}) if
\code{binarize_method} is \code{"count"}.}
}
\value{
A logical vector of binarized mask.
}
\description{
Binarize brain mask
}
