% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{config_file_tracking}
\alias{config_file_tracking}
\title{Configure tracking of neural data files}
\usage{
config_file_tracking(config)
}
\arguments{
\item{config}{A \code{\link[=data.frame]{data.frame()}} storing the specifications of data used.}
}
\value{
Add two columns to \code{config} specifying the file name and the target
name for the data.
}
\description{
Configure tracking of neural data files
}
