% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{config_file_tracking}
\alias{config_file_tracking}
\title{Configure file tracking for neural data}
\usage{
config_file_tracking(
  config,
  dir_neural = "data/neural",
  tar_name_neural = "file_neural",
  name_suffix = ""
)
}
\arguments{
\item{config}{A \code{\link[=data.frame]{data.frame()}} storing the specifications of neural data
used.}

\item{dir_neural}{A character scalar specifying the directory of neural data.}

\item{tar_name_neural}{A character scalar specifying the name of the target
tracking the neural data.}

\item{name_suffix}{A character scalar specifying the name suffix for the
targets.}
}
\value{
A new \code{\link[=data.frame]{data.frame()}} with the \code{file} and \code{tar_neural} fields added.
}
\description{
Configure file tracking for neural data
}
