% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{do_cpm2}
\alias{do_cpm2}
\title{Perform CPM by correctly join neural and behavioral data}
\usage{
do_cpm2(
  neural,
  behav,
  kfolds,
  thresh_method,
  thresh_level,
  bias_correct = TRUE,
  id_cols = "sub_id"
)
}
\arguments{
\item{neural}{A \code{data.frame} storing the neural data.}

\item{behav}{A \code{data.frame} storing the behavioral data.}

\item{kfolds, thresh_method, thresh_level, bias_correct}{See \code{cpm2()}.}

\item{id_cols}{A character vector specifying the column names for subject
identifiers.}
}
\value{
A list of results from \code{cpm2()}.
}
\description{
Perform CPM by correctly join neural and behavioral data
}
