% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{include_g_fitting}
\alias{include_g_fitting}
\title{Fit and extract g factor scores}
\usage{
include_g_fitting(indices, df_ov, include_comp_rel = TRUE)
}
\arguments{
\item{indices}{Raw behavior data.}

\item{df_ov}{A \code{\link[=data.frame]{data.frame()}} specifying observed variables in the field
\code{tasks}.}

\item{include_comp_rel}{A logical value indicating if \code{comp_rel} should be
included. \code{comp_rel} means the composite reliability of latent factors.}
}
\value{
A list of new target objects to fit and extract g factor scores.
}
\description{
This will generate targets used to fit a one-g factor model and predict
factor scores from it.
}
