% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-factory.R
\name{permute_cpm2}
\alias{permute_cpm2}
\title{Target factory for CPM permutation}
\usage{
permute_cpm2(
  behav,
  hypers,
  file_neural_prefix = NULL,
  name_suffix = "",
  subjs_subset = NULL,
  by_brain_mask = NULL,
  split_hyper = NULL,
  subjs_info = NULL,
  batches = 4,
  reps = 5
)
}
\arguments{
\item{behav}{The name for the behavioral data. Should be a symbol.}

\item{hypers}{A \code{\link[=data.frame]{data.frame()}} storing the hyper parameters passed to the
\code{values} argument of \code{\link[tarchetypes:tar_map_rep]{tarchetypes::tar_map_rep()}}. The \code{"modal"},
\code{"parcel"} and \code{"gsr"} fields must be present in \code{hypers} to specify which
neural data to use, and \code{"thresh_method"} and \code{"thresh_level"} are required
to specify threshold method parameters.}

\item{file_neural_prefix}{The name prefix for neural file targets. Must be
the same as that you specified in \code{name_prefix} argument of
\code{\link[=prepare_neural_files]{prepare_neural_files()}}.}

\item{name_suffix}{The name suffix for the target.}

\item{by_brain_mask}{The \code{by} specification for \code{\link[=extract_brain_mask]{extract_brain_mask()}}.
Default is the names matching those of \code{behav} and \code{hypers}, i.e.,
\code{any_of(c(names(behav), names(hypers)))}.}

\item{split_hyper, subjs_info}{If one of these two parameters is specified,
the other must be specified, too. \code{split_hyper} specifies the field used to
split neural data to perform different CPM calculations, e.g., different
gender. Note this field must be present in both \code{hypers} and \code{subjs_info}.}

\item{batches, reps}{The number of batches and repetitions passed to
\code{\link[tarchetypes:tar_map_rep]{tarchetypes::tar_map_rep()}}.}
}
\value{
A new target object to calculate the permutation results.
}
\description{
This will generate batches of CPM permutation for targets to use. Note these
targets depends on those created by \code{\link[=prepare_neural_files]{prepare_neural_files()}}, you should
create those targets first.
}
