% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_mask.R
\name{prepare_adjacency}
\alias{prepare_adjacency}
\title{Prepare adjacency matrix}
\usage{
prepare_adjacency(mask, labels, ...)
}
\arguments{
\item{mask}{A vector of brain mask, should be the upper triangle of the
original full mask.}

\item{labels}{A data frame of ROI labels, should contain \code{label_hemi} column.
Typically the output of \code{\link[=prepare_roi_labels]{prepare_roi_labels()}}.}

\item{...}{Further arguments passed to \code{\link[=binarize_mask]{binarize_mask()}}.}
}
\value{
A data frame of adjacency matrix, with columns \code{row}, \code{col},
\code{degree}, \code{count} and \code{prop}.
}
\description{
Prepare adjacency matrix
}
