% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{prepare_adjacency}
\alias{prepare_adjacency}
\title{Prepare adjacency matrix}
\usage{
prepare_adjacency(mask, ...)
}
\arguments{
\item{mask}{A vector of brain mask, should be the upper triangle of the
original full mask.}

\item{...}{Further arguments passed to \code{\link[=binarize_mask]{binarize_mask()}}.}
}
\value{
A \code{data.frame} with \code{row}, \code{col} and additional values.
}
\description{
Prepare adjacency matrix
}
