% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scoring.R
\name{regress_covariates}
\alias{regress_covariates}
\title{Regress out covariates}
\usage{
regress_covariates(data, subjs_info, covars = NULL, cond = NULL)
}
\arguments{
\item{data}{A data frame with subject identifiers and outcome variables.}

\item{subjs_info}{A data frame with subject identifiers and covariates. The
FD values should be named as \code{mean_fd_task} and \code{mean_fd_rest} for n-back
task and resting state data, respectively. And the first column should be
the subject identifier.}

\item{covars}{A character vector specifying the column names of covariates
to be included. If set as \code{NULL} (default), all covariates will be
included.}

\item{cond}{A character string specifying which FD values to be regressed
out. It can be either \code{nbackrun1}, \code{rest}, or \code{run1rest}. If set as
\code{NULL} (default), no FD values will be regressed out.}
}
\value{
A data frame with residuals.
}
\description{
Regress out covariates
}
