% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cpm.R
\name{regress_covariates}
\alias{regress_covariates}
\title{Regress out covariates}
\usage{
regress_covariates(data, covars, subjs_info)
}
\arguments{
\item{data}{A data frame with subject identifiers and outcome variable.}

\item{covars}{A character vector specifying covariates.}

\item{subjs_info}{A data frame with subject identifiers and covariates.}
}
\value{
A data frame with residuals.
}
\description{
Regress out covariates
}
