% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scoring.R
\name{regress_covariates}
\alias{regress_covariates}
\title{Regress out covariates}
\usage{
regress_covariates(data, subjs_info, covars = TRUE, extracov = NULL)
}
\arguments{
\item{data}{A data frame with subject identifiers and outcome variables.}

\item{subjs_info}{A data frame with subject identifiers and covariates. The
FD values should be named as \code{mean_fd_task} and \code{mean_fd_rest} for n-back
task and resting state data, respectively. And the first column should be
the subject identifier.}

\item{covars}{A character vector specifying the column names of covariates to
be included. If set as \code{TRUE} (default), all covariates will be regressed
out (not including FD values, those are treated in \code{cond}). If set as
\code{NULL}, no covariates will be regressed out.}

\item{extracov}{A numeric value indicating if and which FD values should be
included. \code{0} means no FD values will be included. \code{1} means only FD values
for n-back task will be included. \code{2} means only FD values for resting
state data will be included. \code{NULL} (default) means no FD values will be
included.}
}
\value{
A data frame with residuals.
}
\description{
Regress out covariates
}
