% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{summarise_adjacency}
\alias{summarise_adjacency}
\title{Summarise adjacency matrix}
\usage{
summarise_adjacency(adj_mat, labels)
}
\arguments{
\item{adj_mat}{A matrix of adjacency.}

\item{labels}{A vector of labels. Must have the same length as the number of
rows/columns of the adjacency matrix.}
}
\value{
A data frame of summarised adjacency matrix. It could be converted
as a matrix of upper triangle if \code{x} and \code{y} are used as row and column
respectively.
}
\description{
Summarise adjacency matrix
}
