% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_corrplot}
\alias{visualize_corrplot}
\title{Visualize network as a correlation plot}
\usage{
visualize_corrplot(
  adj_mat,
  model_type,
  labels,
  ...,
  which = c("n", "total", "prop", "enrich"),
  thresh = NULL,
  digits = NULL
)
}
\arguments{
\item{adj_mat}{The adjacency matrix.}

\item{model_type}{A character string specifying the type of the edge. Can be
\code{"pos"} or \code{"neg"}.}

\item{labels}{A vector of labels. Must have the same length as the number of
rows/columns of the adjacency matrix.}

\item{...}{Further arguments passed to \code{\link[corrplot:corrplot]{corrplot::corrplot()}}. These
arguments are not allowed to be changed: \code{method}, \code{type}, \code{is.corr} and
\code{col} which are set to \code{"shade"}, \code{"upper"}, \code{FALSE} and \code{"Reds"} or
\code{"Blues"} depending on the \code{model_type}.}

\item{which}{A character string specifying which value to use.}

\item{thresh}{A numeric value specifying the threshold of the value to be
visualized. If the value is less than the threshold, it will be set as
\code{NA}.}

\item{digits}{A numeric value specifying the number of digits to be rounded
up.}
}
\value{
See \code{\link[corrplot:corrplot]{corrplot::corrplot()}}.
}
\description{
Visualize network as a correlation plot
}
