% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_corrplot}
\alias{visualize_corrplot}
\title{Visualize network as a correlation plot}
\usage{
visualize_corrplot(adj_mat, type, labels, which = "contrib", range = c(0, 100))
}
\arguments{
\item{adj_mat}{The adjacency matrix.}

\item{type}{A character string specifying the type of the edge. Can be
\code{"pos"} or \code{"neg"}.}

\item{labels}{A vector of labels. Must have the same length as the number of
rows/columns of the adjacency matrix.}

\item{which}{A character string specifying which value to use. Can be
\code{"degree"}, \code{"n"}, \code{"relative"} or \code{"contrib"}.}

\item{range}{A numeric vector of length two specifying the range of the
values to be visualized.}
}
\value{
See \code{\link[corrplot:corrplot]{corrplot::corrplot()}}.
}
\description{
Visualize network as a correlation plot
}
