% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize brain mask as chord diagram}
\usage{
visualize_network(
  adj_df,
  roi_labels,
  link_val = NULL,
  link_color = c("white", "black"),
  group_by_hemi = TRUE
)
}
\arguments{
\item{adj_df}{A data frame contains adjacency matrix, should be the upper
triangle of the original full matrix.}

\item{roi_labels}{A data frame contains ROI labels and colors.}

\item{link_val}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variable names
from \code{adj_df} used as the value of links. If \code{NULL}, the link value will be
set as the third column of \code{adj_df}.}

\item{link_color}{A character string of length two, specifying the color of
the lowest and highest link value.}

\item{group_by_hemi}{A logical value indicating whether to group ROIs by
hemisphere.}
}
\value{
Invisible \code{NULL}.
}
\description{
Visualize brain mask as chord diagram
}
